Merged PR 1533: Improve welcome page
Thank you for your contribution to the poc.configuration-management repo.

+semver: minor

Leave the above line in the PR description. This ensures that the patch version will be
bumped whenever merged to the main branch.
For a 'minor' or 'major' release change 'patch' by the corresponding value.

!WARNING! You're submitting a Feature PR

Before submitting this PR, please make sure:

- [X] Your code builds clean without any errors or warnings
- [X] You are using approved terminology
- [X] You have added unit tests
- [X] Special ToDo for Feature branches

Improve welcome page