Merged PR 1583: [devtick-xx] change dir from .md file
Thank you for your contribution to the poc.configuration-management repo.

+semver: patch

Leave the above line in the PR description. This ensures that the patch version will be
bumped whenever merged to the main branch.
For a 'minor' or 'major' release change 'patch' by the corresponding value.

!WARNING! You're submitting a PR to main branch this may affect the productive environment

Before submitting this PR, please make sure:

- [ ] Your code builds clean without any errors or warnings
- [ ] You are using approved terminology
- [ ] You have added unit tests
- [ ] Special ToDo for submitting into main

[devtick-xx] change dir from .md file